Extend ScmUrlResolver, use statics in templates,extract templateContextBuilder
