+++ applications/argocd/nginx/helm-umbrella/values.ftl.yaml
@@ -1,9 +1,11 @@
+<#assign DockerImageParser=statics['com.cloudogu.gitops.utils.DockerImageParser']>
 nginx:
-<#if nginxImage??>
+<#if config.images.nginx?has_content>
+<#assign nginxImageObject = DockerImageParser.parse(config.images.nginx)>
   image:
-    registry: ${nginxImage.registry}
-    repository: ${nginxImage.repository}
-    tag: ${nginxImage.tag}
+    registry: ${nginxImageObject.registry}
+    repository: ${nginxImageObject.repository}
+    tag: ${nginxImageObject.tag}
 </#if>
   <#if config.registry.createImagePullSecrets == true>
