+++ exercises/broken-application/broken-application.ftl.yaml
@@ -1,3 +1,4 @@
+<#assign DockerImageParser=statics['com.cloudogu.gitops.utils.DockerImageParser']>
 apiVersion: apps/v1
 kind: Deploymentz
 metadata:
@@ -18,7 +19,11 @@ spec:
     spec:
       containers:
         - name: broken-application
-          image: <#if nginxImage??>${nginxImage}<#else>bitnami/nginx:1.25.1</#if>
+          image: <#if config.images.nginx?has_content>
+                ${DockerImageParser.parse(config.images.nginx)}
+            <#else>
+                bitnami/nginx:1.25.1
+            </#if>
           ports:
             - containerPort: 8080
 <#if config.application.podResources == true>
