+++ src/main/groovy/com/cloudogu/gitops/features/ExternalSecretsOperator.groovy
@@ -75,7 +75,7 @@ class ExternalSecretsOperator extends Feature implements FeatureWithImage {
                             'Chart.yaml'))['version']
 
             deployer.deployFeature(
-                    ScmUrlResolver.repoUrl(config, repoNamespaceAndName),
+                    ScmUrlResolver.scmmRepoUrl(config, repoNamespaceAndName),
                     "external-secrets",
                     '.',
                     externalSecretsVersion,
