+++ src/main/groovy/com/cloudogu/gitops/features/IngressNginx.groovy
@@ -74,7 +74,7 @@ class IngressNginx extends Feature implements FeatureWithImage {
                             'Chart.yaml'))['version']
 
             deployer.deployFeature(
-                    ScmUrlResolver.repoUrl(config, repoNamespaceAndName),
+                    ScmUrlResolver.scmmRepoUrl(config, repoNamespaceAndName),
                     'ingress-nginx',
                     '.',
                     ingressNginxVersion,
