+++ src/main/groovy/com/cloudogu/gitops/features/Vault.groovy
@@ -133,7 +133,7 @@ class Vault extends Feature implements FeatureWithImage {
                             'Chart.yaml'))['version']
 
             deployer.deployFeature(
-                    ScmUrlResolver.repoUrl(config, repoNamespaceAndName),
+                    ScmUrlResolver.scmmRepoUrl(config, repoNamespaceAndName),
                     'vault',
                     '.',
                     vaultVersion,
